+++ src/test/java/net/sf/jabref/logic/exporter/MSBibExportFormatTestFiles.java
@@ -4,6 +4,7 @@ import java.io.File;
 import java.io.IOException;
 import java.net.URISyntaxException;
 import java.nio.charset.Charset;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -14,13 +15,11 @@ import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import net.sf.jabref.BibDatabaseContext;
-import net.sf.jabref.Globals;
 import net.sf.jabref.logic.importer.ImportFormatPreferences;
 import net.sf.jabref.logic.importer.fileformat.BibtexImporter;
 import net.sf.jabref.model.entry.BibEntry;
 import net.sf.jabref.preferences.JabRefPreferences;
 
-import com.google.common.base.Charsets;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -41,18 +40,17 @@ public class MSBibExportFormatTestFiles {
     public MSBibExportFormat msBibExportFormat;
     public BibtexImporter testImporter;
 
-    public static final String PATH_TO_FILE = "src/test/resources/net/sf/jabref/logic/exporter/";
-
     @Parameter
     public String filename;
+    public Path resourceDir;
 
     @Rule
     public TemporaryFolder testFolder = new TemporaryFolder();
 
 
     @Parameters(name = "{0}")
-    public static Collection<String> fileNames() throws IOException {
-        try (Stream<Path> stream = Files.list(Paths.get(PATH_TO_FILE))) {
+    public static Collection<String> fileNames() throws IOException, URISyntaxException {
+        try (Stream<Path> stream = Files.list(Paths.get(MSBibExportFormatTestFiles.class.getResource("").toURI()))) {
             return stream.map(n -> n.getFileName().toString()).filter(n -> n.endsWith(".bib"))
                     .filter(n -> n.startsWith("MsBib")).collect(Collectors.toList());
         }
@@ -60,25 +58,25 @@ public class MSBibExportFormatTestFiles {
 
     @Before
     public void setUp() throws Exception {
-        Globals.prefs = JabRefPreferences.getInstance();
+        resourceDir = Paths.get(MSBibExportFormatTestFiles.class.getResource("").toURI());
         databaseContext = new BibDatabaseContext();
-        charset = Charsets.UTF_8;
+        charset = StandardCharsets.UTF_8;
         msBibExportFormat = new MSBibExportFormat();
         tempFile = testFolder.newFile();
-        testImporter = new BibtexImporter(ImportFormatPreferences.fromPreferences(Globals.prefs));
+        testImporter = new BibtexImporter(ImportFormatPreferences.fromPreferences(JabRefPreferences.getInstance()));
     }
 
     @Test
-    public final void testPerformExport() throws IOException, URISyntaxException, SaveException {
+    public final void testPerformExport() throws IOException, SaveException {
         String xmlFileName = filename.replace(".bib", ".xml");
-        Path importFile = Paths.get(MSBibExportFormatTestFiles.class.getResource(filename).toURI());
+        Path importFile = resourceDir.resolve(filename);
         String tempFilename = tempFile.getCanonicalPath();
-        List<BibEntry> entries = testImporter.importDatabase(importFile, Charset.defaultCharset()).getDatabase()
+        List<BibEntry> entries = testImporter.importDatabase(importFile, StandardCharsets.UTF_8).getDatabase()
                 .getEntries();
 
         msBibExportFormat.performExport(databaseContext, tempFile.getPath(), charset, entries);
 
-        List<String> expected = Files.readAllLines(Paths.get(PATH_TO_FILE + xmlFileName));
+        List<String> expected = Files.readAllLines(resourceDir.resolve(xmlFileName));
         List<String> exported = Files.readAllLines(Paths.get(tempFilename));
         Collections.sort(expected);
         Collections.sort(exported);