+++ src/test/java/net/sf/jabref/logic/exporter/MsBibExportFormatTest.java
@@ -8,9 +8,7 @@ import java.util.Collections;
 import java.util.List;
 
 import net.sf.jabref.BibDatabaseContext;
-import net.sf.jabref.Globals;
 import net.sf.jabref.model.entry.BibEntry;
-import net.sf.jabref.preferences.JabRefPreferences;
 
 import com.google.common.base.Charsets;
 import org.junit.Before;
@@ -19,7 +17,6 @@ import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 
 import static org.junit.Assert.assertEquals;
-
 
 public class MsBibExportFormatTest {
 
@@ -34,7 +31,6 @@ public class MsBibExportFormatTest {
 
     @Before
     public void setUp() throws Exception {
-        Globals.prefs = JabRefPreferences.getInstance();
         databaseContext = new BibDatabaseContext();
         charset = Charsets.UTF_8;
         msBibExportFormat = new MSBibExportFormat();